Add TEMP_DIR variable and concept.
TEMP_DIR is used when LOG_DIR is not available. The enxt time when both
are available and jit is invoked, all files from TEMP_DIR will be copied
to LOG_DIR. When the copy succeeds the TEMP_DIR files will be deleted.

TEMP_DIR should be a local directory whereas LOG_DIR can be remote.